refactor: split capture worklet module
